feat: add angular form
form folder contains the build files for the angular form.
styles.css and styles.css.map are the main files for the form.
main.js and main.js.map are the main files for the form.
polyfills.js and polyfills.js.map are the polyfills for the form.
index-angular.html is the main file for the form.